# 8331117: [PPC64] secondary_super_cache does not scale well