--- a/src/hotspot/cpu/ppc/macroAssembler_ppc.hpp
+++ b/src/hotspot/cpu/ppc/macroAssembler_ppc.hpp
@@ -606,7 +606,7 @@ class MacroAssembler: public Assembler {
 
   // As above, but with a constant super_klass.
   // The result is in Register result, not the condition codes.
-  bool lookup_secondary_supers_table(Register r_sub_klass,
+  void lookup_secondary_supers_table(Register r_sub_klass,
                                      Register r_super_klass,
                                      Register temp1,
                                      Register temp2,