PPC64 implementation of [JDK-8180450](https://bugs.openjdk.org/browse/JDK-8180450). Please review!
I noticed that `r_array_length` is sometimes 0 and I don't see code for that on x86. Any idea? (This has been addressed in the discussion.)
How can we verify it? By comparing the performance using the micro benchmarks?

Micro benchmark results without patch (measured on Power10 with 2*8 hardware threads):
```
Original 
SecondarySuperCacheHits: 13.033 ±(99.9%) 0.058 ns/op [Average]
SecondarySuperCacheInterContention.test     avgt   15  432.366 ±  8.364  ns/op
SecondarySuperCacheInterContention.test:t1  avgt   15  432.310 ±  8.460  ns/op
SecondarySuperCacheInterContention.test:t2  avgt   15  432.422 ± 10.819  ns/op
SecondarySuperCacheIntraContention.test  avgt   15  355.192 ± 3.597  ns/op
SecondarySupersLookup.testNegative00  avgt   15  12.274 ± 0.026  ns/op
SecondarySupersLookup.testNegative01  avgt   15  12.300 ± 0.039  ns/op
SecondarySupersLookup.testNegative02  avgt   15  12.304 ± 0.034  ns/op
SecondarySupersLookup.testNegative03  avgt   15  12.276 ± 0.050  ns/op
SecondarySupersLookup.testNegative04  avgt   15  12.235 ± 0.044  ns/op
SecondarySupersLookup.testNegative05  avgt   15  12.308 ± 0.156  ns/op
SecondarySupersLookup.testNegative06  avgt   15  12.291 ± 0.048  ns/op
SecondarySupersLookup.testNegative07  avgt   15  12.307 ± 0.052  ns/op
SecondarySupersLookup.testNegative08  avgt   15  12.398 ± 0.075  ns/op
SecondarySupersLookup.testNegative09  avgt   15  12.552 ± 0.122  ns/op
SecondarySupersLookup.testNegative10  avgt   15  12.490 ± 0.083  ns/op
SecondarySupersLookup.testNegative16  avgt   15  12.565 ± 0.092  ns/op
SecondarySupersLookup.testNegative20  avgt   15  19.059 ± 0.958  ns/op
SecondarySupersLookup.testNegative30  avgt   15  19.268 ± 0.124  ns/op
SecondarySupersLookup.testNegative32  avgt   15  20.059 ± 0.114  ns/op
SecondarySupersLookup.testNegative40  avgt   15  25.117 ± 0.368  ns/op
SecondarySupersLookup.testNegative50  avgt   15  32.735 ± 0.359  ns/op
SecondarySupersLookup.testNegative55  avgt   15  34.866 ± 0.152  ns/op
SecondarySupersLookup.testNegative56  avgt   15  35.492 ± 0.276  ns/op
SecondarySupersLookup.testNegative57  avgt   15  36.620 ± 0.334  ns/op
SecondarySupersLookup.testNegative58  avgt   15  37.226 ± 0.180  ns/op
SecondarySupersLookup.testNegative59  avgt   15  37.774 ± 0.241  ns/op
SecondarySupersLookup.testNegative60  avgt   15  38.627 ± 1.451  ns/op
SecondarySupersLookup.testNegative61  avgt   15  39.395 ± 0.249  ns/op
SecondarySupersLookup.testNegative62  avgt   15  40.047 ± 0.377  ns/op
SecondarySupersLookup.testNegative63  avgt   15  40.703 ± 0.416  ns/op
SecondarySupersLookup.testNegative64  avgt   15  41.067 ± 0.216  ns/op
SecondarySupersLookup.testPositive01  avgt   15  12.333 ± 0.037  ns/op
SecondarySupersLookup.testPositive02  avgt   15  12.345 ± 0.042  ns/op
SecondarySupersLookup.testPositive03  avgt   15  12.353 ± 0.075  ns/op
SecondarySupersLookup.testPositive04  avgt   15  12.320 ± 0.021  ns/op
SecondarySupersLookup.testPositive05  avgt   15  12.326 ± 0.046  ns/op
SecondarySupersLookup.testPositive06  avgt   15  12.327 ± 0.063  ns/op
SecondarySupersLookup.testPositive07  avgt   15  12.317 ± 0.040  ns/op
SecondarySupersLookup.testPositive08  avgt   15  12.356 ± 0.042  ns/op
SecondarySupersLookup.testPositive09  avgt   15  12.346 ± 0.044  ns/op
SecondarySupersLookup.testPositive10  avgt   15  12.371 ± 0.040  ns/op
SecondarySupersLookup.testPositive16  avgt   15  12.371 ± 0.078  ns/op
SecondarySupersLookup.testPositive20  avgt   15  12.295 ± 0.048  ns/op
SecondarySupersLookup.testPositive30  avgt   15  12.249 ± 0.170  ns/op
SecondarySupersLookup.testPositive32  avgt   15  12.158 ± 0.026  ns/op
SecondarySupersLookup.testPositive40  avgt   15  12.173 ± 0.049  ns/op
SecondarySupersLookup.testPositive50  avgt   15  12.194 ± 0.053  ns/op
SecondarySupersLookup.testPositive60  avgt   15  12.153 ± 0.032  ns/op
SecondarySupersLookup.testPositive63  avgt   15  12.218 ± 0.088  ns/op
SecondarySupersLookup.testPositive64  avgt   15  12.208 ± 0.038  ns/op
TypePollution.instanceOfInterfaceSwitchLinearNoSCC          avgt   12  11470.410 ± 81.462  ns/op
TypePollution.instanceOfInterfaceSwitchLinearSCC            avgt   12  10784.154 ± 31.282  ns/op
TypePollution.instanceOfInterfaceSwitchTableNoSCC           avgt   12  11455.959 ± 76.254  ns/op
TypePollution.instanceOfInterfaceSwitchTableSCC             avgt   12  10723.693 ± 32.825  ns/op
TypePollution.parallelInstanceOfInterfaceSwitchLinearNoSCC  avgt   12    152.832 ±  1.474  ms/op
TypePollution.parallelInstanceOfInterfaceSwitchLinearSCC    avgt   12    132.049 ±  4.752  ms/op
TypePollution.parallelInstanceOfInterfaceSwitchTableNoSCC   avgt   12    150.466 ±  7.390  ms/op
TypePollution.parallelInstanceOfInterfaceSwitchTableSCC     avgt   12    134.714 ±  1.128  ms/op
```

With patch:
```
SecondarySuperCacheHits.test  avgt   15  13.020 ± 0.006  ns/op
SecondarySuperCacheInterContention.test     avgt   15  423.927 ± 11.187  ns/op
SecondarySuperCacheInterContention.test:t1  avgt   15  427.683 ± 12.014  ns/op
SecondarySuperCacheInterContention.test:t2  avgt   15  420.171 ± 14.093  ns/op
SecondarySuperCacheIntraContention.test  avgt   15  355.758 ± 6.324  ns/op
SecondarySupersLookup.testNegative00  avgt   15  12.192 ± 0.035  ns/op
SecondarySupersLookup.testNegative01  avgt   15  12.218 ± 0.067  ns/op
SecondarySupersLookup.testNegative02  avgt   15  12.191 ± 0.030  ns/op
SecondarySupersLookup.testNegative03  avgt   15  12.196 ± 0.033  ns/op
SecondarySupersLookup.testNegative04  avgt   15  12.213 ± 0.042  ns/op
SecondarySupersLookup.testNegative05  avgt   15  12.173 ± 0.018  ns/op
SecondarySupersLookup.testNegative06  avgt   15  12.197 ± 0.040  ns/op
SecondarySupersLookup.testNegative07  avgt   15  12.178 ± 0.030  ns/op
SecondarySupersLookup.testNegative08  avgt   15  12.197 ± 0.059  ns/op
SecondarySupersLookup.testNegative09  avgt   15  12.159 ± 0.047  ns/op
SecondarySupersLookup.testNegative10  avgt   15  12.157 ± 0.038  ns/op
SecondarySupersLookup.testNegative16  avgt   15  12.190 ± 0.050  ns/op
SecondarySupersLookup.testNegative20  avgt   15  12.178 ± 0.101  ns/op
SecondarySupersLookup.testNegative30  avgt   15  12.228 ± 0.062  ns/op
SecondarySupersLookup.testNegative32  avgt   15  12.196 ± 0.063  ns/op
SecondarySupersLookup.testNegative40  avgt   15  12.203 ± 0.074  ns/op
SecondarySupersLookup.testNegative50  avgt   15  12.177 ± 0.056  ns/op
SecondarySupersLookup.testNegative55  avgt   15  12.212 ± 0.065  ns/op
SecondarySupersLookup.testNegative56  avgt   15  12.205 ± 0.063  ns/op
SecondarySupersLookup.testNegative57  avgt   15  12.215 ± 0.061  ns/op
SecondarySupersLookup.testNegative58  avgt   15  12.196 ± 0.042  ns/op
SecondarySupersLookup.testNegative59  avgt   15  12.211 ± 0.055  ns/op
SecondarySupersLookup.testNegative60  avgt   15  12.409 ± 0.181  ns/op
SecondarySupersLookup.testNegative61  avgt   15  12.445 ± 0.184  ns/op
SecondarySupersLookup.testNegative62  avgt   15  12.461 ± 0.207  ns/op
SecondarySupersLookup.testNegative63  avgt   15  59.441 ± 0.549  ns/op
SecondarySupersLookup.testNegative64  avgt   15  59.834 ± 0.312  ns/op
SecondarySupersLookup.testPositive01  avgt   15  12.176 ± 0.020  ns/op
SecondarySupersLookup.testPositive02  avgt   15  12.188 ± 0.036  ns/op
SecondarySupersLookup.testPositive03  avgt   15  12.207 ± 0.042  ns/op
SecondarySupersLookup.testPositive04  avgt   15  12.214 ± 0.041  ns/op
SecondarySupersLookup.testPositive05  avgt   15  12.239 ± 0.032  ns/op
SecondarySupersLookup.testPositive06  avgt   15  12.193 ± 0.054  ns/op
SecondarySupersLookup.testPositive07  avgt   15  12.182 ± 0.039  ns/op
SecondarySupersLookup.testPositive08  avgt   15  12.190 ± 0.020  ns/op
SecondarySupersLookup.testPositive09  avgt   15  12.195 ± 0.065  ns/op
SecondarySupersLookup.testPositive10  avgt   15  12.178 ± 0.029  ns/op
SecondarySupersLookup.testPositive16  avgt   15  12.202 ± 0.033  ns/op
SecondarySupersLookup.testPositive20  avgt   15  12.192 ± 0.048  ns/op
SecondarySupersLookup.testPositive30  avgt   15  12.150 ± 0.034  ns/op
SecondarySupersLookup.testPositive32  avgt   15  12.141 ± 0.030  ns/op
SecondarySupersLookup.testPositive40  avgt   15  12.159 ± 0.063  ns/op
SecondarySupersLookup.testPositive50  avgt   15  12.174 ± 0.052  ns/op
SecondarySupersLookup.testPositive60  avgt   15  12.178 ± 0.046  ns/op
SecondarySupersLookup.testPositive63  avgt   15  12.200 ± 0.060  ns/op
SecondarySupersLookup.testPositive64  avgt   15  12.179 ± 0.041  ns/op
TypePollution.instanceOfInterfaceSwitchLinearNoSCC          avgt   12  11458.792 ±  31.853  ns/op
TypePollution.instanceOfInterfaceSwitchLinearSCC            avgt   12  10816.996 ± 117.688  ns/op
TypePollution.instanceOfInterfaceSwitchTableNoSCC           avgt   12   7409.874 ±  10.493  ns/op
TypePollution.instanceOfInterfaceSwitchTableSCC             avgt   12   7378.197 ±  10.471  ns/op
TypePollution.parallelInstanceOfInterfaceSwitchLinearNoSCC  avgt   12    153.648 ±   2.088  ms/op
TypePollution.parallelInstanceOfInterfaceSwitchLinearSCC    avgt    9    136.728 ±   1.059  ms/op
TypePollution.parallelInstanceOfInterfaceSwitchTableNoSCC   avgt    8     74.106 ±   0.272  ms/op
TypePollution.parallelInstanceOfInterfaceSwitchTableSCC     avgt    5     73.969 ±   0.528  ms/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8331117](https://bugs.openjdk.org/browse/JDK-8331117): [PPC64] secondary_super_cache does not scale well (**Bug** - P3)


### Reviewers
 * [Richard Reingruber](https://openjdk.org/census#rrich) (@reinrich - **Reviewer**) ⚠️ Review applies to [bea2f938](https://git.openjdk.org/jdk/pull/19368/files/bea2f938bce50ecbea81ef1a0b48176155f25ee5)
 * [Amit Kumar](https://openjdk.org/census#amitkumar) (@offamitkumar - Committer)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/19368/head:pull/19368` \
`$ git checkout pull/19368`

Update a local copy of the PR: \
`$ git checkout pull/19368` \
`$ git pull https://git.openjdk.org/jdk.git pull/19368/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 19368`

View PR using the GUI difftool: \
`$ git pr show -t 19368`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/19368.diff">https://git.openjdk.org/jdk/pull/19368.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/19368#issuecomment-2127239578)