@TheRealMDoerr 2024-05-24
@theRealAph: It would be great if you could take a look and see if you can spot any bug. Especially, I wonder why `r_array_length` happens to be 0 in some cases, but x86 doesn't check.