@theRealAph 2024-05-27
> PPC64 implementation of [JDK-8180450](https://bugs.openjdk.org/browse/JDK-8180450). Please review! I noticed that `r_array_length` is sometimes 0 and I don't see code for that on x86. Any idea? How can we verify it? By comparing the performance using the micro benchmarks?

Run all of tier1 with `-XX:+VerifySecondarySupers`