@TheRealMDoerr 2024-05-27
> > @theRealAph: It would be great if you could take a look and see if you can spot any bug. Especially, I wonder why `r_array_length` happens to be 0 in some cases, but x86 doesn't check.
> 
> Why would it not be zero? Some classes don't have secondary super types. In addition, 12ns is very slow. I don't understand that.

I had to check for `r_array_length  >= 0` here: https://github.com/openjdk/jdk/pull/19368/files#diff-0f708565c9e138b8013165540634368334f5d1df2ba437e39696e9791440050dR2312
The x86 implementation doesn't do that and I wonder why. Doesn't it access stale memory (https://github.com/openjdk/jdk/blob/be1d374bc54d43aae3b3c1feace22d38fe2156b6/src/hotspot/cpu/x86/macroAssembler_x86.cpp#L4967)?