@theRealAph 2024-05-27
`cmpdi(CCR0, r_bitmap, (bit + 1) & Klass::SECONDARY_SUPERS_TABLE_MASK);`

Why is this a compare, not a bit test?