@TheRealMDoerr 2024-05-28
Thank you so much for finding my bug! This explains why I got array_lenght 0. Fixed and added assertion (see 2nd commit).