@theRealAph 2024-05-29
> Performance seems to be not affected by that bug. 

That is extremely suspicious.