@offamitkumar 2024-06-07
@TheRealMDoerr I got one test failure on PPC with these changes: 
```diff
diff --git a/src/hotspot/share/runtime/globals.hpp b/src/hotspot/share/runtime/globals.hpp
index 6bfb260606b..70897a1066e 100644
--- a/src/hotspot/share/runtime/globals.hpp
+++ b/src/hotspot/share/runtime/globals.hpp
@@ -1988,13 +1988,13 @@ const int ObjectAlignmentInBytes = 8;
                 "rewriting/transformation independently of the JVMTI "      \
                 "can_{retransform/redefine}_classes capabilities.")         \
                                                                             \
-  product(bool, UseSecondarySupersCache, true, DIAGNOSTIC,                  \
+  product(bool, UseSecondarySupersCache, false, DIAGNOSTIC,                  \
                 "Use secondary supers cache during subtype checks.")        \
                                                                             \
-  product(bool, UseSecondarySupersTable, false, DIAGNOSTIC,                 \
+  product(bool, UseSecondarySupersTable, true, DIAGNOSTIC,                 \
                 "Use hash table to lookup secondary supers.")               \
                                                                             \
-  product(bool, VerifySecondarySupers, false, DIAGNOSTIC,                   \
+  product(bool, VerifySecondarySupers, true, DIAGNOSTIC,                   \
           "Check that linear and hashed secondary lookups return the same result.") \
                                                                             \
   product(bool, StressSecondarySupers, false, DIAGNOSTIC,                   \
```

```
==============================
Test summary
==============================
   TEST                                              TOTAL  PASS  FAIL ERROR   
   jtreg:./test/hotspot/jtreg/compiler/c2/irTests/ProfileAtTypeCheck.java
>>                                                       1     0     1     0 <<
==============================
TEST FAILURE
```


But if I revert the changes I had done, then it passes. Same situation I'm facing on s390x. Is this expected ? 

failure log: 
[type_profile_failure.log](https://github.com/user-attachments/files/15741205/type_profile_failure.log)