@TheRealMDoerr 2024-06-14
Thanks for reviewing! Your suggestions are all good (see my update).