@TheRealMDoerr 2024-06-14
Thanks for the review!
I've tried some very simple statistics to check that the first lookup directly hits as expected: java -Xcomp -XX:-TieredCompilation -version
secondary supers direct hit ratio: 537 / 539 (the other 2 ones were hit with the first check in the first slow path loop)