@TheRealMDoerr 2024-06-15
I've added a check for Power7 because [JDK-8331859](https://bugs.openjdk.org/browse/JDK-8331859) is not yet implemented. It should get removed by that one again. I think we should have this check in case this PR gets backported.