@theRealAph 2024-06-15
> That doesn't look like a platform specific thing. I'm getting the same result on x86_64. @theRealAph: Is that a known limitation or is it worth a new JBS issue?

I've never seen this. It must be a regression. I'll have a look.