@reinrich 2024-06-17
> > That doesn't look like a platform specific thing. I'm getting the same result on x86_64. @theRealAph: Is that a known limitation or is it worth a new JBS issue?
> 
> Ah, I see. The test is doing some IR node counts for Klass loads, and `-UseSecondarySupersCache` deletes one of those loads. So it's not actually a behavioural change. I'm not sure what to do about this.

I think the `@IR` rule can be duplicated using `applyIf` to distinguish between `-XX:-UseSecondarySupersCache` and `-XX:+UseSecondarySupersCache`